refactor: extract ensureBufCommand to shared utility in utils.ts
